+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 - Fix regression in Slider component
 - Fix do not set 'aria-hidden' if element has already been removed from the accessibility tree
+- Fix Switcher component does not ignore disabled toggles
 
 ## 3.5.16 (December 11, 2020)